+++ examples/windows/blank_window.cc
@@ -4,7 +4,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//      http://www.apache.org/licenses/LICENSE-2.0
+//      https://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,19 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "vscge/core/types.h"
+#include <vscge/vscge.h>
 
-#include <Windows.h>
-
-namespace vs {
-  Size:
-constexpr operator COORD() {
-  return {(SHORT)width, (SHORT)height};
-
+class BlankWindow : public vs::Application {
+  void OnUpdate(const vs::Timestep&) override {}
 };
-Rect:
-  constexpr operator SMALL_RECT() const {
-    return {(SHORT)x, (SHORT)y, (SHORT)(x + width - 1),
-            (SHORT)(y + height - 1)};
-  };
-}  // namespace vs
+
+VS_REGISTER_APP(BlankWindow);