+++ vscge/include/vscge/event/event.h
@@ -17,6 +17,7 @@
 
 #include <string>
 
+#include "vscge/core/api.h"
 #include "vscge/core/core.h"
 
 namespace vs {
@@ -34,6 +35,7 @@ enum class EventType {
 
 // TODO(Victor): See if it makes sense to add a `handled` to Event.
 struct VS_API Event {
+  virtual ~Event()                     = default;
   virtual EventType Type() const       = 0;
   virtual std::string TypeName() const = 0;
 };