+++ vscge/src/core/application.cc
@@ -34,52 +34,30 @@
 #pragma warning(disable : 4100)
 
 namespace vs {
+Application::Application(const Size& screen_size, const Size& font_size) {
+  window_.Initialize();
 #if 0
-Application::Application(const Size& screen_size, const Size& font_size)
-    : buffer_in_(GetStdHandle(STD_INPUT_HANDLE)) {
-  SetConsoleMode(buffer_in_, ENABLE_EXTENDED_FLAGS | ENABLE_WINDOW_INPUT |
-                                 ENABLE_MOUSE_INPUT);
-
   Renderer::Initialize(GetStdHandle(STD_OUTPUT_HANDLE), screen_size,
   font_size);
+#endif
   Logger::Initialize();
-
-#if 0
-  SetConsoleCtrlHandler(
-      static_cast<PHANDLER_ROUTINE>(Application::CloseHandler), true);
-#endif
-
-  is_running_ = true;
 }
-#endif
 
-#if 0
-BOOL Application::CloseHandler(DWORD) {
-  Logger::Log("Stopping application!");
-  is_running_ = false;
-  std::unique_lock lock(closing_);
-  has_finished_.wait(lock);
-  return true;
-}
-#endif
-
-void Application::Start() {
+void Application::Initialize() {
   VS_PROFILE_BEGIN_SESSION("Start", "runtime.json");
   OnStart();
 
   std::thread game_loop(VS_BIND_THREAD(Application::MainLoop));
-  std::thread event_loop(VS_BIND_THREAD(Application::EventListener));
   game_loop.join();
-  event_loop.join();
   VS_PROFILE_END_SESSION();
 
-  has_finished_.notify_all();
+  Window::has_finished_.notify_all();
 }
 
 void Application::MainLoop() {
   Stopwatch timer;
   timer.Start();
-  while (is_running_) {
+  while (Window::is_running_) {
     VS_PROFILE_FUNCTION();
     Timestep timestep = timer.Stop();
     std::string timing_message =
@@ -92,112 +70,6 @@ void Application::MainLoop() {
     OnUpdate(timestep);
 
     // Renderer::Render();
-  }
-}
-
-// TODO(Victor): Move to a platform layer.
-void Application::EventListener() {
-  Ref<MouseEvent> previous_mouse_event;
-  std::unordered_map<Key, Ref<KeyEvent>> previous_key_event;
-  while (is_running_) {
-    DWORD num_events;
-    GetNumberOfConsoleInputEvents(buffer_in_, &num_events);
-
-    constexpr size_t kMaxEvents                       = 32;
-    std::array<INPUT_RECORD, kMaxEvents> event_buffer = {};
-    ReadConsoleInput(buffer_in_, event_buffer.data(), num_events, &num_events);
-    for (size_t i = 0; i < num_events; ++i) {
-      auto read_event = event_buffer.at(i);
-      switch (read_event.EventType) {
-        case KEY_EVENT: {
-          VS_PROFILE_SCOPE("KEY_EVENT");
-          KEY_EVENT_RECORD record = read_event.Event.KeyEvent;  // NOLINT
-          Key key                 = static_cast<Key>(record.wVirtualKeyCode);
-          bool is_down            = record.bKeyDown;
-
-          // INFO(Victor): for some reason, this can't be an std::string_view,
-          // should investigate more.
-          std::string message = "Received key " +
-                                std::to_string(static_cast<int>(key)) + ' ' +
-                                (is_down ? "down" : "up");
-          Logger::Log(message, Logger::Level::kCore);
-
-          // If first time, create a dummy KeyReleasedEvent
-          if (!previous_key_event[key]) {
-            previous_key_event[key] =
-                CreateRef<KeyReleasedEvent>(KeyReleasedEvent{key});
-          }
-
-          message = "Previous state: " + previous_key_event[key]->TypeName();
-          Logger::Log(message, Logger::Level::kCore);
-
-          if (previous_key_event[key]->Type() != EventType::kKeyReleased &&
-              !is_down) {
-            KeyReleasedEvent event{key};
-            previous_key_event[key] = CreateRef<KeyReleasedEvent>(event);
-          } else if (previous_key_event[key]->Type() ==
-                         EventType::kKeyReleased &&
-                     is_down) {
-            KeyPressedEvent event{key};
-            previous_key_event[key] = CreateRef<KeyPressedEvent>(event);
-          } else {
-            // TODO(Victor): Add an event KeyHold.
-            previous_key_event[key] = CreateRef<KeyEvent>(KeyEvent{key});
-          }
-
-          OnEvent(previous_key_event[key]);
-
-          break;
-        }
-        case MOUSE_EVENT: {
-          VS_PROFILE_SCOPE("MOUSE_EVENT");
-          MOUSE_EVENT_RECORD record = read_event.Event.MouseEvent;  // NOLINT
-          Point position = {record.dwMousePosition.X, record.dwMousePosition.Y};
-          MouseButtons buttons;
-
-          if (record.dwButtonState & FROM_LEFT_1ST_BUTTON_PRESSED) {
-            buttons.left = true;
-          }
-          if (record.dwButtonState & RIGHTMOST_BUTTON_PRESSED) {
-            buttons.right = true;
-          }
-
-          // If first time, create a dummy MouseEvent
-          if (!previous_mouse_event) {
-            previous_mouse_event =
-                CreateRef<MouseEvent>(MouseEvent{{0, 0}, {false, false}});
-          }
-
-          // TODO(Victor): The logic to get button presses/releases is mostly
-          // certainly wrong.
-          if (record.dwEventFlags & MOUSE_MOVED) {
-            // NOLINTNEXTLINE
-            MouseMovedEvent event = {previous_mouse_event->position, position,
-                                     buttons};
-            previous_mouse_event  = CreateRef<MouseMovedEvent>(event);
-          } else {
-            if ((previous_mouse_event->buttons.left && !buttons.left) ||
-                (previous_mouse_event->buttons.right && !buttons.right)) {
-              MouseButtonReleasedEvent event = {position, buttons};
-              previous_mouse_event = CreateRef<MouseButtonReleasedEvent>(event);
-            } else if ((!previous_mouse_event->buttons.left && buttons.left) ||
-                       (!previous_mouse_event->buttons.right &&
-                        buttons.right)) {
-              MouseButtonPressedEvent event = {position, buttons};
-              previous_mouse_event = CreateRef<MouseButtonPressedEvent>(event);
-            }
-          }
-
-          OnEvent(previous_mouse_event);
-
-          break;
-        }
-        case WINDOW_BUFFER_SIZE_EVENT:
-        case MENU_EVENT:
-        case FOCUS_EVENT:
-          break;
-      }
-    }
   }
 }
 }  // namespace vs